Raise lower version bounds on base to >= 4.9 (GHC 8.0+)
This is needed in a subsequent commit which imports `Data.Kind`, which was
first introduced in `base-4.9.0.0`.

Now that `base-4.9` or later is required, the `base-compat` dependency is no
longer required, so this patch removes it as well. I also removed some
`MIN_VERSION_base` CPP that is no longer useful.