Replace binary-parsers dependency with attoparsec
`binary-parsers` does not build with GHC 9.2, and it is unclear if the library
is actively maintained anymore. The only thing we are using `binary-parsers`
for is to have parser combinators over lazy `ByteString`s, but one can just as
well accomplish the same thing with `attoparsec`. This patch replaces
`binary-parsers` with `attoparsec`.

Fixes #10.