Replace binary-parsers dependency with attoparsec's Zepto
`binary-parsers` does not build with GHC 9.2, and it is unclear if the library
is actively maintained anymore. The only thing we are using `binary-parsers`
for is to have parser combinators over `ByteString`s, but one can just as
well accomplish the same thing with `attoparsec`.

This patch replaces `binary-parsers` with `attoparsec`. In particular, it uses
the `Data.Attoparsec.Zepto` module, which provides an especially
high-performance parser that outperforms `binary-parsers` on a large test case.
While `ZeptoT` is not performant for parsers that have a lot of backtracking or
recursion, it is fortunately the case that neither happen that often with the
AIGER file format.

Fixes #10.